requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.1.0
pymongo==4.8.0
pytz==2024.2
redis==5.0.8
requests==2.32.3
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.11.1
aiogram==2.25.1
aiohttp==3.8.6
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.1.0
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
requests==2.32.3
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
yarl==1.11.1
